convert back to float

+++ src/aabb.h
@@ -1,5 +1,5 @@
 #ifndef __AABB__
 #define __AABB__
 
- typedef double aabb[2][3];
+ typedef float aabb[2][3];
 #endif
